Using checksum calculation to avoid applying the Deployments again and again on azure
